Add Board, ActionSpace and Space classes and refactor existing ones
Board is meant to be initialized when the game is started. It contains vector of all
of the spaces needed for the game.Space class is parent class for all of the
spaces on the board. All of the spaces which invoke some kind of action from user
(other than unowned properties) are members of the ActionSpace class.

Remove 'using namespace std' from all of the .hpp files and add action index to Card class.

If header file with 'using namespace std' is included, it forces
(often involuntary) usage of the std namespace which can cause naming conflicts
and non-obvious ambiguities.Action index will be used in the future to identify
the action that is invoked when particular card is drawn.