Add Game class, update Makefile and Player class
Game class will be the main controller for the game, it will control
player turns, initialize board and players, throw dice, control auctions
and probably move players in an infinite loop. Constructor of this class
takes number of players which should be provided via qt button in the future.

Changed some of the functions so that program will compile.

Added new class into the Makefile.