Add load/save menu
Menu for loading from/saving to .xml files currently only supports
loading. When user chooses to load from file, file browser dialog is
shown prompting user to chose file from which to load saved game.
Another dialog is shown after specifing a file asking user if he
wishes to continue with loading because current game will be lost. There
are three options available: yes,no and save(save option still not functional).

LoadXML class is in charge of parsing .xml file and creating a new game
with players and their owned spaces based on data provided in .xml file.
There is still work to be done in area of bad file reading (alert
message will be shown, and loading will be aborted). This will probably
be accomplished via signal function(s?).

setIsOnMortgage function is added to all of the ownable classes in order
to allow LoadXML class to set mortgage.

Game class nextPlayer method is now based on position in a vector of
players rather than on player's id. This was necessary in order to be
able to show player tabs correctly. It is possible that player's id
doesn't correspond to his play order (i.e. game could be saved at the
moment in which player with id 3 is supposed to play, so in .xml file
his id will be 3, but his turn will be 0).

mainwindow class is altered in order to correctly (for now :S) incorporate
new features. Some minor refactoring was done in this class. QMenu and
QAction private variables were added and implemented. There is still
work to be done for save functionality. Some off_by_one errors when
accessing vector of players were corrected.