Add save functionality
It is now possible to save current state of the game to the specified
.xml file. That same file can later be loaded and game can be continued.

When loading from file, user is prompted wheter he/she wants to save current
game before loading and appropriate actions are taken based on the
user's answer. There was only load functionality in previous version.

Rename LoadXML class to HandleXML: same class handles loading and
saving.

Add some kind of bad file detection: if any of the elements is empty,
abort loading and report it to the mainwindow.

TODO: disallow anyone but the game itself to alter saved .xml files.