Fix load, add sell house button and xml test file
There was a problem when loading from file because board didn't have
information about space owners. Spaces are now added to the board when they
are read from file.

New game constructor that accepts vector of Player pointers, Bank and
Board pointers.

Sell house button is only visible when player has viable houses to sell
on selected space.If he chooses to sell the house, house is
removed from the space and it is "returned" to the bank.