Add delegate to the table view, add new methods to space class and its subclasses
Implemented delegate for the table view, that does all the work for the graphic presentation in the cells.
Players should now be visible as colored circles, and if there are buildings built on a certain space,
they should be visible as grey squares. In order for the delegate to work properly,
few methods have been added to space class and its subclasses, that handle which (and if any) player is positioned on that space.